+++ quarkus/src/test/java/org/fuin/cqrs4j/quarkus/view/QuarkusViewManagerTest.java
@@ -7,7 +7,7 @@ import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.junit.QuarkusTestProfile;
 import io.quarkus.test.junit.TestProfile;
 import jakarta.inject.Inject;
-import org.fuin.cqrs4j.core.JpaView;
+import org.fuin.cqrs4j.core.View;
 import org.fuin.cqrs4j.esc.ProjectionService;
 import org.fuin.esc.api.EventStore;
 import org.fuin.esc.api.ProjectionAdminEventStore;
@@ -18,7 +18,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.function.Consumer;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.fail;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
@@ -26,21 +26,21 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 /**
- * Test for the {@link QuarkusJpaViewManager} class.
+ * Test for the {@link QuarkusViewManager} class.
  */
 @Disabled("TODO Implement!")
 @QuarkusTest
-@TestProfile(QuarkusJpaViewManagerTest.class)
-class QuarkusJpaViewManagerTest implements QuarkusTestProfile {
+@TestProfile(QuarkusViewManagerTest.class)
+class QuarkusViewManagerTest implements QuarkusTestProfile {
 
     @Inject
-    QuarkusJpaViewManager testee;
+    QuarkusViewManager testee;
 
     @InjectMock
     Scheduler scheduler;
 
     @InjectMock
-    List<JpaView> rawViews;
+    List<View> rawViews;
 
     @InjectMock
     EventStore eventstore;
@@ -60,12 +60,12 @@ class QuarkusJpaViewManagerTest implements QuarkusTestProfile {
     void createViews() {
 
         // GIVEN
-        final JpaView view = mock(JpaView.class);
-        final List<JpaView> views = List.of(view);
+        final View view = mock(View.class);
+        final List<View> views = List.of(view);
         when(rawViews.stream()).thenReturn(views.stream());
 
         final Scheduler.JobDefinition jobDefinition = mock(Scheduler.JobDefinition.class);
-        when(scheduler.newJob(view.getName())).thenReturn(jobDefinition);
+        when(scheduler.newJob(view.getBeanName())).thenReturn(jobDefinition);
 
         // WHEN
         testee.createViews();
@@ -74,8 +74,6 @@ class QuarkusJpaViewManagerTest implements QuarkusTestProfile {
         verify(jobDefinition, times(1)).setCron(view.getCron());
         verify(jobDefinition, times(1)).setTask((Consumer<ScheduledExecution>) any());
         verify(jobDefinition, times(1)).schedule();
-
-
 
     }