+++ springboot/src/main/java/org/fuin/cqrs4j/springboot/view/SpringViewRegistry.java
@@ -1,0 +1,50 @@
+package org.fuin.cqrs4j.springboot.view;
+
+import jakarta.annotation.Nonnull;
+import org.fuin.cqrs4j.core.SimpleViewRegistry;
+import org.fuin.cqrs4j.core.View;
+import org.fuin.cqrs4j.core.ViewRegistry;
+import org.springframework.beans.factory.config.BeanDefinition;
+import org.springframework.beans.factory.config.ConfigurableBeanFactory;
+
+import java.util.List;
+
+import static org.springframework.beans.factory.config.BeanDefinition.SCOPE_PROTOTYPE;
+
+/**
+ * Registry that verifies that views have 'prototype' scope.
+ */
+public class SpringViewRegistry implements ViewRegistry {
+
+    private final SimpleViewRegistry delegate;
+
+    public SpringViewRegistry(ConfigurableBeanFactory beanFactory, List<View> views) {
+        // Verify that all beans are dependent scoped
+        for (final View view : views) {
+            final BeanDefinition beanDefinition = beanFactory.getMergedBeanDefinition(view.getBeanName());
+            final String scope = beanDefinition.getScope();
+            if (!SCOPE_PROTOTYPE.equals(scope)) {
+                throw new IllegalStateException("Bean scope must be @Scope(SCOPE_PROTOTYPE), but was @Scope("
+                        + scope + ") for bean: " + view.getBeanClass().getName());
+            }
+        }
+        delegate = new SimpleViewRegistry(views);
+    }
+
+    @Nonnull
+    @Override
+    public List<Entry> getViews() {
+        return delegate.getViews();
+    }
+
+    @Override
+    public int size() {
+        return delegate.size();
+    }
+
+    @Override
+    public boolean isEmpty() {
+        return delegate.isEmpty();
+    }
+
+}