+++ test/quarkus/src/main/java/org/fuin/cqrs4j/quarkus/test/model/AbstractPersonsView.java
@@ -1,7 +1,7 @@
 package org.fuin.cqrs4j.quarkus.test.model;
 
 import jakarta.persistence.EntityManager;
-import org.fuin.cqrs4j.core.JpaView;
+import org.fuin.cqrs4j.core.View;
 import org.fuin.ddd4j.core.Event;
 import org.fuin.ddd4j.core.EventType;
 import org.slf4j.Logger;
@@ -13,13 +13,14 @@ import java.util.Set;
 /**
  * Handles the events required to maintain the persons view.
  */
-public abstract class AbstractPersonsView implements JpaView {
+public abstract class AbstractPersonsView implements View {
 
     private static final Logger LOG = LoggerFactory.getLogger(AbstractPersonsView.class);
 
-    @Override
-    public String getName() {
-        return "persons-view";
+    private final EntityManager em;
+
+    protected AbstractPersonsView(EntityManager em) {
+        this.em = em;
     }
 
     @Override
@@ -28,17 +29,17 @@ public abstract class AbstractPersonsView implements JpaView {
     }
 
     @Override
-    public void handleEvents(final EntityManager em, final List<Event> events) {
+    public void handleEvents(final List<Event> events) {
         for (final Event event : events) {
             if (event instanceof PersonCreatedEvent ev) {
-                handlePersonCreatedEvent(em, ev);
+                handlePersonCreatedEvent(ev);
             } else {
                 throw new IllegalStateException("Cannot handle event: " + event);
             }
         }
     }
 
-    private void handlePersonCreatedEvent(final EntityManager em, final PersonCreatedEvent event) {
+    private void handlePersonCreatedEvent(final PersonCreatedEvent event) {
         LOG.info("Handle {}: {}", event.getClass().getSimpleName(), event);
         final PersonEntity entity = em.find(PersonEntity.class, event.getId().asBaseType());
         if (entity == null) {