+++ test/springboot/src/main/java/org/fuin/cqrs4j/springboot/test/view/PersonsView.java
@@ -1,24 +1,43 @@
 package org.fuin.cqrs4j.springboot.test.view;
 
 import jakarta.persistence.EntityManager;
-import org.fuin.cqrs4j.core.JpaView;
+import org.fuin.cqrs4j.core.View;
 import org.fuin.cqrs4j.springboot.test.model.PersonCreatedEvent;
 import org.fuin.cqrs4j.springboot.test.model.PersonEntity;
 import org.fuin.ddd4j.core.Event;
 import org.fuin.ddd4j.core.EventType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.context.annotation.Scope;
+import org.springframework.stereotype.Component;
 
 import java.util.List;
 import java.util.Set;
 
-public class PersonsView implements JpaView {
+import static org.springframework.beans.factory.config.BeanDefinition.SCOPE_PROTOTYPE;
+
+@Component(PersonsView.BEAN_NAME)
+@Scope(SCOPE_PROTOTYPE)
+public class PersonsView implements View {
 
     private static final Logger LOG = LoggerFactory.getLogger(PersonsView.class);
 
+    public static final String BEAN_NAME = "persons-view";
+
+    private final EntityManager em;
+
+    public PersonsView(final EntityManager em) {
+        this.em = em;
+    }
+
     @Override
-    public String getName() {
-        return "persons-view";
+    public String getBeanName() {
+        return BEAN_NAME;
+    }
+
+    @Override
+    public Class<? extends View> getBeanClass() {
+        return PersonsView.class;
     }
 
     @Override
@@ -33,17 +52,17 @@ public class PersonsView implements JpaView {
     }
 
     @Override
-    public void handleEvents(final EntityManager em, final List<Event> events) {
+    public void handleEvents(final List<Event> events) {
         for (final Event event : events) {
             if (event instanceof PersonCreatedEvent ev) {
-                handlePersonCreatedEvent(em, ev);
+                handlePersonCreatedEvent(ev);
             } else {
                 throw new IllegalStateException("Cannot handle event: " + event);
             }
         }
     }
 
-    private void handlePersonCreatedEvent(final EntityManager em, final PersonCreatedEvent event) {
+    private void handlePersonCreatedEvent(final PersonCreatedEvent event) {
         LOG.info("Handle {}: {}", event.getClass().getSimpleName(), event);
         final PersonEntity entity = em.find(PersonEntity.class, event.getId().asBaseType());
         if (entity == null) {